pyproject.toml
[tool.poetry]
name = "chatui"
version = "0.1.0"
description = "A Streamlit app for communicating with GPT by OpenAI."
authors = ["Fabian <dev@fabiannie.haus>"]
readme = "README.md"
packages = [{include = "chatui", from = "src"}]

[tool.poetry.dependencies]
python =  "~3.11"
numpy = "^1.24.2"
pandas = "^1.5.3"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
streamlit = "^1.27.2"
langchain = {extras = ["docarray"], version = "^0.0.327"}
pypdf = "^3.16.4"
sentence-transformers = "^2.2.2"
chromadb = "^0.4.15"
tiktoken = "^0.5.1"
rapidocr-onnxruntime = "^1.3.8"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
jupyter = "^1.0.0"
black = "^23.1.0"
jupyter-black = "^0.3.1"
mypy = "^1.1.1"
isort = "^5.12.0"
pytest = "^7.2.2"
pytest-mock = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
